BF+TST - round integers before writing to file
Fixes some nasty rounding errors when converting between large and small
integers.

If we are writing to an integer type, do explicit rounding before
writing, because the 'astype' method rounds down and this can make quite
a mess with large scaling factors.

Change from in-place modification of arrays to write; it will raise an
error in the next version of numpy when we mean to cast to float, and
it's more explicit.

Some minor refactorings to array_to_file to try and increase speed, but
the benchmarks are pretty flat.

Expanded tests of array_to_file to cover the areas of the optimizations.

Expanded tests of array_to_file and back again for integer min and max
values.