RF+BF - struggle with type conversions and floats
Working rig of testing low-level conversions between large and small
integers, and between floats and integers.

Upcasting floats for writing the data causes some slow-down of the
writing process as measured with the benchmarks - in the order of 10
percent I think.  There is probably room to relax the width of float to
use for the casting.