NF - commit can be untestable for unexpected error
If we are using ``error-txt`` to look for an error, and the match fails,
and the nosetests run returns a non-zero exit code, that means one or
more tests failed, but not in the way we were expecting.  With new
``error-untest`` option, this can identify the commit as untestable;
without, the commit is presumed good (and the failure is presumed
irrelevant).