RF: use WrapStruct and tests for Ecat main header
Inherit from WrapStruct, remove duplicate code, use generic WrapStruct
tests.

Wrapstruct does not by default force the returned results of __getitem__
to be scalar.  We therefore have to force the scaling factors from the
header to be scalars to allow the scalefactors to upcast the data array
to floating point.