RF: expand scaling checks in get_slope_inter
Make header `get_slope_inter()` do checks for non-finite etc scaling for
all Analyze types.

Raise error in `get_slope_inter` if there is a valid slope but invalid
intercept.  This is the default when reading a header from disc from the
log checks, but I'm going to remove the log checks soon, and I want to
preserve the logic set in `TestAnalyzeImage.test_header_scaling()`