RF+API: no finite thresholds for no-scale writer
ArrayWriter can't use scaling, but had a generic 'scaling_needed'
function that assumed that arrays of all nans, infs will get set to
zero, because of the behavior of the ``finite_range`` method.

This makes sense if we are scaling, because then it is most sensible to
scale to the largest / smallest non-finite value.  For no-scaling, it
makes more sense to clip any values outside the dtype range to the max /
min.

Implement this and test.

This is a very minor API change, because ArrayWriter until a few commits
ago raised an error for any float to integer conversion that wasn't an
array just composed of non-finite values.  So, the only change of
effective API is that arrays of non-finite values passes to ArrayWriter
now raise a WriterError by default, instead of being quietly set to all
zero.